Issue with pagination YOU will get 5$
I have an issue with this pagination you will get rewarded with 5$.
There is a 3 to 4 sub-category when u click on it category this page (https://laalmon.com/category.php) will open but the issue is at the bottom of the page. At the bottom, there is a pagination option 2,3,4, but when I clicked to see page 2 it did not work. On the other hand, when u click on ALL PRODUCT OPTION (bottom of the page) the pagination works perfectly.

category.php pagination 2,3,4,5 NOT working.
all_products.php pagination 2,3,4,5 WORKING.